Refactored shopping session and cart items endpoints and add withProps to stores

+++ apps/api/src/cart-items/cart-items.service.ts
@@ -2,28 +2,34 @@ import {
   ForbiddenException,
   Injectable,
   NotFoundException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { CreateCartItemDto } from './dto/create-cart-item.dto';
 import { UpdateCartItemDto } from './dto/update-cart-item.dto';
 import { PrismaService } from '../prisma/prisma.service';
-import { ShoppingSessionsService } from '../shopping-sessions/shopping-sessions.service';
 import { decode } from 'jsonwebtoken';
 import { Book, ShoppingSession } from '@prisma/client';
+import { ShoppingSessionEntity } from '../shopping-sessions/entities/shopping-session.entity';
 
 @Injectable()
 export class CartItemsService {
-  constructor(
-    private prisma: PrismaService,
-    private shoppingSessionService: ShoppingSessionsService,
-  ) {}
+  constructor(private prisma: PrismaService) {}
 
   async create(authHeader: string, { bookId, quantity }: CreateCartItemDto) {
     const userId = String(decode(authHeader.split(' ')[1]).sub);
+
+    if (!userId) {
+      throw new UnauthorizedException('You should log in');
+    }
 
     const shoppingSession = await this.prisma.shoppingSession.findUnique({
       where: { userId },
       select: { id: true },
     });
+
+    if (!shoppingSession) {
+      throw new NotFoundException('Not found shopping session');
+    }
 
     const existingCartItem = await this.prisma.cartItem.findUnique({
       where: {
@@ -89,7 +95,13 @@ export class CartItemsService {
           },
         }));
 
-      this.shoppingSessionService.updateTotal(cartItem.shoppingSessionId);
+      try {
+        this.#updateTotal(cartItem.shoppingSessionId);
+      } catch (error) {
+        throw new Error(
+          error ?? 'Error occurred while updating total price of cart items',
+        );
+      }
 
       return cartItem;
     }
@@ -110,7 +122,13 @@ export class CartItemsService {
       },
     });
 
-    this.shoppingSessionService.updateTotal(cartItem.shoppingSessionId);
+    try {
+      this.#updateTotal(cartItem.shoppingSessionId);
+    } catch (error) {
+      throw new Error(
+        error ?? 'Error occurred while updating total price of cart items',
+      );
+    }
 
     return {
       ...cartItem,
@@ -121,11 +139,17 @@ export class CartItemsService {
     };
   }
 
-  findOne(shoppingSessionId: ShoppingSession['id'], bookId: Book['id']) {
-    return this.prisma.cartItem.findUnique({
+  async findOne(shoppingSessionId: ShoppingSession['id'], bookId: Book['id']) {
+    const cartItem = await this.prisma.cartItem.findUnique({
       where: { bookId_shoppingSessionId: { bookId, shoppingSessionId } },
       include: { shoppingSession: { select: { userId: true } } },
     });
+
+    if (!cartItem) {
+      throw new NotFoundException('Not found cart item');
+    }
+
+    return cartItem;
   }
 
   async update(
@@ -135,6 +159,10 @@ export class CartItemsService {
     data: UpdateCartItemDto,
   ) {
     const userId = String(decode(authHeader.split(' ')[1]).sub);
+
+    if (!userId) {
+      throw new UnauthorizedException('You should log in');
+    }
 
     const cartItem = await this.findOne(shoppingSessionId, bookId);
 
@@ -170,7 +198,13 @@ export class CartItemsService {
       },
     });
 
-    this.shoppingSessionService.updateTotal(updatedCartItem.shoppingSessionId);
+    try {
+      this.#updateTotal(cartItem.shoppingSessionId);
+    } catch (error) {
+      throw new Error(
+        error ?? 'Error occurred while updating total price of cart items',
+      );
+    }
 
     return {
       ...updatedCartItem,
@@ -187,6 +221,10 @@ export class CartItemsService {
     bookId: Book['id'],
   ) {
     const userId = String(decode(authHeader.split(' ')[1]).sub);
+
+    if (!userId) {
+      throw new UnauthorizedException('You should log in');
+    }
 
     const deletedCartItem = await this.prisma.cartItem.delete({
       where: {
@@ -206,7 +244,17 @@ export class CartItemsService {
       },
     });
 
-    this.shoppingSessionService.updateTotal(deletedCartItem.shoppingSessionId);
+    if (!deletedCartItem) {
+      throw new NotFoundException('Not found cart item');
+    }
+
+    try {
+      this.#updateTotal(deletedCartItem.shoppingSessionId);
+    } catch (error) {
+      throw new Error(
+        error ?? 'Error occurred while updating total price of cart items',
+      );
+    }
 
     return {
       ...deletedCartItem,
@@ -215,5 +263,32 @@ export class CartItemsService {
         authors: deletedCartItem.book.authors.map((a) => a.author),
       },
     };
+  }
+
+  async #updateTotal(shoppingSessionId: ShoppingSessionEntity['id']) {
+    const shoppingSession = await this.prisma.shoppingSession.findUnique({
+      where: { id: shoppingSessionId },
+      include: {
+        cartItems: {
+          include: {
+            book: true,
+          },
+        },
+      },
+    });
+
+    if (!shoppingSession) {
+      throw new NotFoundException('Not found shopping session');
+    }
+
+    return this.prisma.shoppingSession.update({
+      where: { id: shoppingSessionId },
+      data: {
+        total: shoppingSession.cartItems.reduce(
+          (acc, ct) => acc + ct.book.price * ct.quantity,
+          0,
+        ),
+      },
+    });
   }
 }
